:truck: api: move to-be-published-session-serializer from lib to src
