+++ api/src/certification/session-management/infrastructure/serializers/to-be-published-session-serializer.js
